From remote theme to local files
Previously I was using gem for local, and remote theme for production (GitHub pages). However I need few edits to the theme files (specially to make DT R-html-widget make work) and maybe in future.
So moving on to using all theme files locally. Try to keep changes to theme core files at minimum, so as to limit dependency on the theme.